Implemented updateEmployee method using DTOs
Refactored the updateEmployee method in the EmployeeService interface and its implementation (EmployeeServiceImpl) to utilize Data Transfer Objects (DTOs). Updated the EmployeeMapper interface to correctly map DTOs to entity objects and vice versa. Ensured that the updateEmployee method does not alter the identifier of the Employee entity, resolving potential issues with the entity's identifier being changed unintentionally.