Added population of role_id field when creating a user
In the createUser method of the UserServiceImpl class, population of the role_id field has been added when creating a user. Now, when saving a new user, a corresponding entry in the role table is also created, linked to their role.

Additionally, proper configuration of the @ManyToOne and @JoinColumn annotations has been added in the User entity to establish the relationship with the Role entity.