Added transaction isolation levels in UserServiceImpl and DepartmentServiceImpl
In the UserServiceImpl and DepartmentServiceImpl classes, @Transactional annotations with specified transaction isolation levels were added to methods. Methods that read data from the database are annotated with READ_COMMITTED isolation level, while methods that modify data are annotated with SERIALIZABLE isolation level. This was done to ensure the correct operation of services in a multi-threaded environment, taking into account the requirements for data isolation.

Additionally, @JsonIgnore annotations were added to certain fields in the UserInfo and Authority entities. This was done to exclude information about related User entities from JSON serialization in these classes, as this information will already be included in the serialization of the User entity. This improves the structure of JSON data and reduces redundancy.