Added a test for the employee creation controller
Created a `test` directory with the first test for the `createEmployee` controller. In the test, an `EmployeeCreateDto` object with test data is created, which is then converted to JSON. A POST request is sent to `/employee/create` with the data in JSON format. After receiving the response, the returned JSON is converted to an `EmployeeAfterCreationDto` object. Checked the status of the response (expected 200).